# dist/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "FundooHR Backend API's Specification",
        "description": "FundooHR is not a Software product, it is a thinking which make your task easy and efficient. And we are plotting our thining in software product. Which will make our work easy. Basically we are planning to make product based on some different technologies.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Bridgelabz.com",
            "email": "admin@bridgelabz.com"
        }
    },
    "basePath": "/",
...
    "schemes": [
      "http",
      "https"
    ],
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json",
        "application/xml"
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "description": "It authenticates the user email from company database and it returns success response with email and token. For every REST call you need to pass token and crossponding email with some extra request body which is based on REST call's.",
                "operationId": "for the URL",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [{
                    "name": "emailId",
                    "in": "formData",
                    "description": "Provide a valid email address to authenticate. e.g. your@some-domain.com",
                    "required": true,
                    "type": "string",
                    "default": "admin@bridgelabz.com"
                }, {
                    "name": "password",
                    "in": "formData",
                    "description": "Provide a valid Password",
                    "required": true,
                    "type": "string",
                    "format": "password",
                    "default": "Bridge@123"
                }],
                "responses": {
                    "200": {
                        "description": "Authentcation Successful",
                        "schema": {
                            "$ref": "#/definitions/loginResponse"
                        }
                    }
                }
            }
        },
        "/readDashboardData": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "description": "It return you Dashboard Data",
                "operationId": "findById",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "timeStamp",
                    "in": "query",
                    "description": "timestamp is wrapping the date and time in milliseconds format 1483401600000 ",
                    "required": true,
                    "type": "integer"
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "String"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/DashResponse"
                        }
                    }
                }
            }
        },
        "/readUnmarkedAttendanceEmployee": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "description": "It return the list of employee whoever is not marked attendace for passed date",
                "operationId": "findById",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "timeStamp",
                    "in": "query",
                    "description": "timestamp is wrapping the date and time in milliseconds format 1483401600000 ",
                    "required": true,
                    "type": "integer"
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "String"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/RUAEResponse"
                        }
                    }
                }
            }
        },
        "/readFalloutAttendanceEmployee": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "description": "It return the list of employee whoever is not marked attendace for 3 or more than 3 days",
                "operationId": "findById",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "timeStamp",
                    "in": "query",
                    "description": "timestamp is wrapping the date and time in milliseconds format 1483401600000 ",
                    "required": true,
                    "type": "integer"
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "String"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/RFAEResponse"
                        }
                    }
                }
            }
        },
        "/readLeaveEmployee": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "description": "It return the list of employee who taken 2 or more than 2 days leave",
                "operationId": "findById",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "timeStamp",
                    "in": "query",
                    "description": "timestamp is wrapping the date and time in milliseconds format 1483401600000 ",
                    "required": true,
                    "type": "integer"
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "String"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/RLEResponse"
                        }
                    }
                }
            }
        },
        "/readMonthlyAttendanceSummary": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "description": "It will give you day wise attendance ",
                "operationId": "findById",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "timeStamp",
                    "in": "query",
                    "description": "timestamp is wrapping the date and time in milliseconds format 1483401600000 ",
                    "required": true,
                    "type": "integer"
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "String"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/RMAEResponse"
                        }
                    }
                }
            }
        },
        "/sendEmailToUnmarkedEmployee": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "description": "It sends the email to unmarked employee for passed date(timestamp) ",
                "operationId": "for the URL",
                "produces": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [{
                    "name": "timeStamp",
                    "in": "body",
                    "description": "timestamp is wrapping the date and time in milliseconds format 1483401600000 ",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/SETUESchema"

                    }
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/SETUEResponse"
                        }
                    }
                }
            }
        },
        "/sendEmailToFalloutEmployee": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "description": "It sends the email to Fallout employee for passed month(timestamp) ",
                "operationId": "for the URL",
                "produces": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [{
                    "name": "timeStamp",
                    "in": "body",
                    "description": "timestamp is wrapping the date and time in milliseconds format 1483401600000 ",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/SETFESchema"

                    }
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/SETUFEResponse"
                        }
                    }
                }
            }
        },
        "/sendEmailToLeaveEmployee": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "description": "It sends the email who taken 2 or more than two days leave for passed month(timestamp) ",
                "operationId": "for the URL",
                "produces": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [{
                    "name": "timeStamp",
                    "in": "body",
                    "description": "timestamp is wrapping the date and time in milliseconds format 1483401600000 ",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/SETFESchema"

                    }
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/SETUFEResponse"
                        }
                    }
                }
            }
        },
        "/searchEmployeeByName": {
            "get": {
                "tags": [
                    "Engineer","Search"
                ],
                "description": "It return the employee list with basic details",
                "operationId": "for the URL",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/SEBNResponse"
                        }
                    }
                }
            }
        },
        "/searchEmployee/{searchKey}/{cursor}": {
            "get": {
                "tags": [
                    "Engineer","Search"
                ],
                "description": "It return the employee list with basic details",
                "operationId": "for the URL",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                  {
                      "name": "searchKey",
                      "in": "path",
                      "description": "Here your will pass The seaching key eg,a.",
                      "required": true,
                      "type": "string"
                  },
                  {
                      "name": "cursor",
                      "in": "path",
                      "description": "here you will pass the cursor key default is 0 next is retrieving from previous response",
                      "default":0,
                      "required": true,
                      "type": "string"
                  },
                  {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/SEBNResponse"
                        }
                    }
                }
            }
        },
        "/readEmployeeMonthlyAttendance": {
            "get": {
                "tags": [
                    "Attendance","Engineer"
                ],
                "description": "It return the employee monthly attendace",
                "operationId": "for the URL",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "engineerId",
                    "in": "query",
                    "description": "Here you need to pass Engineer Id",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "timeStamp",
                    "in": "query",
                    "description": "Here you need to pass month,year(timeStamp) for findinf empoyee attendace",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/REMAResponse"
                        }
                    }
                }
            }
        },
        "/createEmployeeDayAttendance": {
            "post": {
                "tags": [
                    "Attendance","Engineer"
                ],
                "description": "It is use to create employee attendace",
                "operationId": "for the URL",
                "produces": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [{
                    "name": "Scehma",
                    "in": "body",
                    "description": "Here you need to pass The json following the Schema",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/CEDASchema"
                    }
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/CEDAResponse"
                        }
                    }
                }
            }
        },
        "/readEmployeeHRData": {
            "get": {
                "tags": [
                    "Engineer"
                ],
                "description": "It is use to retrieve the employee HR data",
                "operationId": "for the URL",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "engineerId",
                    "in": "query",
                    "description": "Here you need to pass The json following the Schema",
                    "required": true,
                    "type": "string",
                    "default": "40002EI"
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/REHDResponse"
                        }
                    }
                }
            }
        },
        "/readEmployeePersonalData": {
            "get": {
                "tags": [
                    "Engineer"
                ],
                "description": "It is use to retrieve the employee Personal data",
                "operationId": "for the URL",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "engineerId",
                    "in": "query",
                    "description": "Here you need to pass The json following the Schema",
                    "required": true,
                    "type": "string",
                    "default": "40002EI"
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/REPDResponse"
                        }
                    }
                }
            }
        },
        "/readEmployeeProfileData": {
            "get": {
                "tags": [
                    "Engineer"
                ],
                "description": "It is use to retrieve the employee Profile (Educational) data",
                "operationId": "for the URL",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "engineerId",
                    "in": "query",
                    "description": "Here you need to pass The json following the Schema",
                    "required": true,
                    "type": "string",
                    "default": "40002EI"
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/REPrDResponse"
                        }
                    }
                }
            }
        },
        "/readEmployeeTrackingData": {
            "get": {
                "tags": [
                    "Engineer"
                ],
                "description": "It is use to retrieve the employee Track data",
                "operationId": "for the URL",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "engineerId",
                    "in": "query",
                    "description": "Here you need to pass The json following the Schema",
                    "required": true,
                    "type": "string",
                    "default": "40002EI"
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/RETDResponse"
                        }
                    }
                }
            }
        },
        "/readEmployeeBankData": {
            "get": {
                "tags": [
                    "Engineer"
                ],
                "description": "It is use to retrieve the employee Bank data",
                "operationId": "for the URL",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name": "engineerId",
                    "in": "query",
                    "description": "Here you need to pass The json following the Schema",
                    "required": true,
                    "type": "string",
                    "default": "40002EI"
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/REBDResponse"
                        }
                    }
                }
            }
        },
        "/updateEmployeeHRData": {
            "put": {
                "tags": [
                    "Engineer"
                ],
                "description": "It is use to Update the employee HR data",
                "operationId": "for the URL",
                "produces": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [{
                    "name": "engineerData",
                    "in": "body",
                    "description": "Here you need to pass The json following the Schema",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/UEHDSchema"
                    }
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/UEHDResponse"
                        }
                    }
                }
            }
        },
        "/updateEmployeePersonalData": {
            "put": {
                "tags": [
                    "Engineer"
                ],
                "description": "It is use to Update the employee Personal data",
                "operationId": "for the URL",
                "produces": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [{
                    "name": "engineerData",
                    "in": "body",
                    "description": "Here you need to pass The json following the Schema",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/UEPDSchema"
                    }
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/UEHDResponse"
                        }
                    }
                }
            }
        },
        "/updateEmployeeProfileData":{
            "put": {
                "tags": [
                    "Engineer"
                ],
                "description": "It is use to Update the employee Profile(Educational) data",
                "operationId": "for the URL",
                "produces": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [{
                    "name": "engineerData",
                    "in": "body",
                    "description": "Here you need to pass The json following the Schema",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/UEPrDSchema"
                    }
                }, {
                    "name": "x-token",
                    "in": "header",
                    "description": "You need to pass token along with your request.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "response",
                        "schema": {
                            "$ref": "#/definitions/UEHDResponse"
                        }
                    }
                }
          }
      },
      "/updateEmployeeTrackingData":{
        "put":{
          "tags": [
              "Engineer"
          ],
          "description": "It is use to Update the employee Tracking data",
          "operationId": "for the URL",
          "produces": [
              "application/x-www-form-urlencoded"
          ],
          "parameters": [{
              "name": "engineerData",
              "in": "body",
              "description": "Here you need to pass The json following the Schema",
              "required": true,
              "schema": {
                  "$ref": "#/definitions/UETDSchema"
              }
          }, {
              "name": "x-token",
              "in": "header",
              "description": "You need to pass token along with your request.",
              "required": true,
              "type": "string"
          }],
          "responses": {
              "200": {
                  "description": "response",
                  "schema": {
                      "$ref": "#/definitions/UEHDResponse"
                  }
              }
          }
        }
      },
      "/updateEmployeeBankData":{
        "put":{
          "tags": [
              "Engineer"
          ],
          "description": "It is use to Update the employee Bank data",
          "operationId": "for the URL",
          "produces": [
              "application/x-www-form-urlencoded"
          ],
          "parameters": [{
              "name": "engineerData",
              "in": "body",
              "description": "Here you need to pass The json following the Schema",
              "required": true,
              "schema": {
                  "$ref": "#/definitions/UEBDSchema"
              }
          }, {
              "name": "x-token",
              "in": "header",
              "description": "You need to pass token along with your request.",
              "required": true,
              "type": "string"
          }],
          "responses": {
              "200": {
                  "description": "response",
                  "schema": {
                      "$ref": "#/definitions/UEHDResponse"
                  }
              }
          }
        }
      },
      "/readAllEmployee":{
        "get":{
          "tags": [
              "Engineer","Report"
          ],
          "description": "It is use to retrieve all Employee Data to generate the Salary report or salary invoice",
          "operationId": "for the URL",
          "produces": [
              "application/x-www-form-urlencoded"
          ],
          "parameters": [{
              "name": "x-token",
              "in": "header",
              "description": "You need to pass token along with your request.",
              "required": true,
              "type": "string"
          }],
          "responses": {
              "200": {
                  "description": "response",
                  "schema": {
                      "$ref": "#/definitions/RAEResponse"
                  }
              }
          }
        }
      },
...
      },
        "definitions": {
            "loginResponse": {
                "properties": {
                    "token": {
                        "required": true,
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhZG1pbkBicmlkZ2VsYWJ6LmNvbSIsImlhdCI6MTQ4Nzc2MDY3NSwiZXhwIjoxNDg4OTcwMjc1fQ.ALOhIMuqtczsrZkR-vtTbceP5djAQNk7CCwoZzKkOCY"
                    },
                    "status": {
                        "required": true,
                        "type": "Interger",
                        "example": "200"
                    },
                    "message": {
                        "required": true,
                        "type": "string",
                        "example": "login success"
                    }
                }
            },
...
            "SETUESchema": {
                "properties": {
                    "timeStamp": {
                        "type": "integer",
                        "required": true,
                        "example": 1483401600000
                    }
                }
            },
...
            "SETFESchema": {
                "properties": {
                    "timeStamp": {
                        "type": "integer",
                        "required": true,
                        "example": 1483401600000
                    }
                }
            },
...
    }
  }